Added Record form.
