+++ scripts/components/EditForm.js
@@ -1,8 +1,6 @@
 import React, { Component } from "react";
-import { Link } from "react-router";
-import Form from "react-jsonschema-form";
 
-import Spinner from "./Spinner";
+import RecordForm from "./RecordForm";
 import { omit } from "../utils";
 
 
@@ -11,29 +9,24 @@ function recordToFormData(record) {
 }
 
 export default class EditForm extends Component {
-  onSubmit = ({formData}) => {
+  onSubmit = (record) => {
     const {params, updateRecord} = this.props;
     const {bid, cid, rid} = params;
-    updateRecord(bid, cid, rid, formData);
+    updateRecord(bid, cid, rid, record);
   }
 
   render() {
     const {params, collection, record} = this.props;
-    const {schema, uiSchema, busy} = collection;
-    const {bid, cid, rid} = params;
+    const {bid, cid} = params;
     return (
       <div>
-        <h1>{bid}/{cid}/{rid}</h1>
-        {busy ? <Spinner /> :
-          <Form
-            formData={recordToFormData(record)}
-            schema={schema}
-            uiSchema={uiSchema}
-            onSubmit={this.onSubmit}>
-            <input type="submit" className="btn btn-primary" value="Update" />
-            {" or "}
-            <Link to={`/buckets/${bid}/collections/${cid}`}>Cancel</Link>
-          </Form>}
+        <h1>Add a new record in {bid}/{cid}</h1>
+        <RecordForm
+          bid={bid}
+          cid={cid}
+          collection={collection}
+          record={recordToFormData(record)}
+          onSubmit={this.onSubmit} />
       </div>
     );
   }
